fix the abs(long) issues with ESMC Field tests on pgi

+++ src/Infrastructure/Field/tests/ESMC_FieldGridGridRegridCsrvUTest.C
@@ -1,4 +1,4 @@
-// $Id: ESMC_FieldGridGridRegridCsrvUTest.C,v 1.2 2012/04/13 16:32:21 rokuingh Exp $
+// $Id: ESMC_FieldGridGridRegridCsrvUTest.C,v 1.3 2012/04/16 16:06:07 rokuingh Exp $
 //
 // Earth System Modeling Framework
 // Copyright 2002-2012, University Corporation for Atmospheric Research, 
@@ -665,7 +665,7 @@ int main(void){
 #ifdef masking
       tol = 100;
 #endif
-      if ( abs((double)( dstfieldptr[p]-exact) ) > tol) {
+      if ( abs((long)(dstfieldptr[p]-exact)) > tol) {
         printf("dstfieldptr [%f,%f]:\n%f /= %f\n", 
                x, y, dstfieldptr[p], exact);
         correct=false;
@@ -674,7 +674,7 @@ int main(void){
     }
   }
   // check that the mass is conserved
-  //if (abs(srcmass - dstmass) > .0001) correct = false;
+  //if (abs((long)(srcmass - dstmass)) > .0001) correct = false;
   //printf("srcmass = %f, dstmass = %f, srcmass-dstmass = %f\n", srcmass, dstmass, srcmass-dstmass);
 
   ESMC_Test((correct==true), name, failMsg, &result, __FILE__, __LINE__, 0);
