+++ src/Infrastructure/Field/tests/ESMC_FieldGridGridRegridUTest.C
@@ -1,4 +1,4 @@
-// $Id: ESMC_FieldGridGridRegridUTest.C,v 1.3 2012/04/13 16:32:21 rokuingh Exp $
+// $Id: ESMC_FieldGridGridRegridUTest.C,v 1.4 2012/04/16 16:06:07 rokuingh Exp $
 //
 // Earth System Modeling Framework
 // Copyright 2002-2012, University Corporation for Atmospheric Research, 
@@ -442,7 +442,7 @@ int main(void){
 #ifdef masking
       tol = 100;
 #endif
-      if ( abs((double)( dstfieldptr[p]-exact) ) > tol) {
+      if ( abs((long)(dstfieldptr[p]-exact)) > tol) {
         printf("dstfieldptr [%f,%f]:\n%f /= %f\n", 
                x, y, dstfieldptr[p], exact);
         correct=false;
