+++ src/Infrastructure/Field/tests/ESMC_FieldGridRegridCsrv2UTest.C
@@ -1,4 +1,4 @@
-// $Id: ESMC_FieldGridRegridCsrv2UTest.C,v 1.3 2012/04/13 16:32:21 rokuingh Exp $
+// $Id: ESMC_FieldGridRegridCsrv2UTest.C,v 1.4 2012/04/16 16:06:07 rokuingh Exp $
 //
 // Earth System Modeling Framework
 // Copyright 2002-2012, University Corporation for Atmospheric Research, 
@@ -618,7 +618,7 @@ int main(void){
 #ifdef nomasking
       tol = .0001;
 #endif
-      if ( abs((double)( dstfieldptr[p]-exact) ) > tol) {
+      if ( abs((long)(dstfieldptr[p]-exact)) > tol) {
         printf("dstfield [%f,%f]:\n%f /= %f\n", 
                x, y , dstfieldptr[p], exact);
         correct=false;
@@ -627,7 +627,7 @@ int main(void){
     }
   }
   // check that the mass is conserved
-  if (abs(srcmass - dstmass) > .0001) correct = false;
+  if (abs((long)(srcmass - dstmass)) > .0001) correct = false;
   //printf("srcmass = %f, dstmass = %f\n", srcmass, dstmass);
 
   ESMC_Test((correct==true), name, failMsg, &result, __FILE__, __LINE__, 0);
